+++ src/main/scala/scorex/api/http/leasing/SignedLeaseRequest.scala
@@ -12,7 +12,7 @@ case class SignedLeaseRequest(@ApiModelProperty(value = "Base58 encoded sender p
                               @ApiModelProperty(required = true)
                               fee: Long,
                               @ApiModelProperty(required = true)
-                              untilBlock: Long,
+                              untilBlock: Int,
                               @ApiModelProperty(value = "Recipient address", required = true)
                               recipient: String,
                               @ApiModelProperty(required = true)
@@ -25,7 +25,7 @@ object SignedLeaseRequest {
     (JsPath \ "sender").read[String] and
       (JsPath \ "amount").read[Long] and
       (JsPath \ "fee").read[Long] and
-      (JsPath \ "untilBlock").read[Long] and
+      (JsPath \ "untilBlock").read[Int] and
       (JsPath \ "recipient").read[String] and
       (JsPath \ "timestamp").read[Long] and
       (JsPath \ "signature").read[String](SignatureReads)
